Move internal modules into separate files.
